Forms with Tight Binding and Reactive forms
Earlier form and list of post/article made to work with tight object type Article (post.ts) also made required changes in the form. On the other hand we learnt the reactive form and the required classes to make it work. Currently displaying of Validation message is not done but we handled dynamic display of messages using the templates.